+++ R/labels.R
@@ -3,11 +3,17 @@
 #' These functions allow you to extract the labels from a
 #' \code{\link{tskrr}} object. The function \code{labels} and the
 #' function \code{dimnames} are aliases and do the exact same
-#' thing.
+#' thing. The functions \code{rownames} and \code{colnames} work like
+#' you would expect. Note that contrary to the latter two, \code{labels}
+#' will never return \code{NULL}. If no labels are found, it will construct
+#' labels using the prefixes defined in the argument \code{prefix}.
 #'
-#' The functions \code{colnames} and \code{rownames} work like you would
-#' expect, with one difference: the rownames or colnames for a \code{tskrr}
-#' object can never be \code{NULL}.
+#' @section Warning:
+#' If the original data didn't contain row- or column names for the
+#' adjacency matrix, \code{rownames} and \code{colnames} will return
+#' \code{NULL}. Other functions will extract the automatically generated
+#' labels, so don't count on \code{rownames} and \code{colnames} if you
+#' want to predict output from other functions!
 #'
 #' @param x a \code{\link{tskrr}} object
 #' @param object a \code{\link{tskrr}} object
@@ -101,7 +107,7 @@ setMethod("rownames",
 setMethod("colnames",
           "tskrr",
           function(x, do.NULL = TRUE, prefix = "col"){
-            rn <- x@labels$g
+            rn <- if(is_homogenous(x)) x@labels$k else  x@labels$g
 
             nolabels <- length(rn) == 1 && is.na(rn)